move cloud/pkg/devicecontroller/apis to cloud/pkg/apis
