+++ cloud/hack/update-codegen.sh
@@ -21,6 +21,6 @@ set -o pipefail
 SCRIPT_ROOT=$(unset CDPATH && cd $(dirname "${BASH_SOURCE[0]}")/../.. && pwd)
 
 ${SCRIPT_ROOT}/vendor/k8s.io/code-generator/generate-groups.sh "deepcopy" \
-github.com/kubeedge/kubeedge/cloud/pkg/client github.com/kubeedge/kubeedge/cloud/pkg/devicecontroller/apis \
+github.com/kubeedge/kubeedge/cloud/pkg/client github.com/kubeedge/kubeedge/cloud/pkg/apis \
 "devices:v1alpha1" \
 --go-header-file ${SCRIPT_ROOT}/cloud/hack/boilerplate/boilerplate.txt
