+++ cloud/pkg/apis/devices/v1alpha1/zz_generated.deepcopy.go
